+++ app/src/main/java/woowacourse/payments/ui/addcard/Password.kt
@@ -1,0 +1,30 @@
+package woowacourse.payments.ui.addcard
+
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import woowacourse.payments.R
+import woowacourse.payments.domain.Password
+
+@Composable
+fun Password(
+    password: Password,
+    modifier: Modifier = Modifier,
+    onValueChange: (String) -> Unit = {},
+) {
+    OutlinedTextField(
+        value = password.toString(),
+        onValueChange = onValueChange,
+        modifier = modifier,
+        placeholder = { Text(stringResource(R.string.add_card_password_placeholder_text)) },
+        label = { Text(stringResource(R.string.add_card_password_label_text)) },
+        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
+        visualTransformation = PasswordVisualTransformation(),
+        singleLine = true,
+    )
+}
