+++ app/src/main/java/woowacourse/payments/ui/addcard/PaymentCard.kt
@@ -1,0 +1,39 @@
+package woowacourse.payments.ui.addcard
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun PaymentCard(modifier: Modifier = Modifier) {
+    Box(
+        contentAlignment = Alignment.CenterStart,
+        modifier =
+            modifier
+                .shadow(8.dp)
+                .size(width = 208.dp, height = 124.dp)
+                .background(
+                    color = Color(0xFF333333),
+                    shape = RoundedCornerShape(5.dp),
+                ),
+    ) {
+        Box(
+            modifier =
+                Modifier
+                    .padding(start = 14.dp, bottom = 10.dp)
+                    .size(width = 40.dp, height = 26.dp)
+                    .background(
+                        color = Color(0xFFCBBA64),
+                        shape = RoundedCornerShape(4.dp),
+                    ),
+        )
+    }
+}
